qmmf-sdk: maintain PCM capture data flow on Stop
Since StopSession is a blocking call and (for PCM
capture) the EOS needs to be propagated to the
QMMF client; the integrity of the data flow needs
to be held until the StopSession call returns.

This was not happening for various PCM capture
use cases.  Made changes to allow the data flow
until the EOS is sent for each track.

Change-Id: I468b8f901c8c67dad84c39f8c4aba01645672a88