recorder: Bypass frame skip logic if StopSession requested
In case of low fps streams, many consecutive frames may need
to be skipped; plus if there are inherent delays in frames received
from camera, this may result in timeout in the thread waiting to
receive frames from Camera. Hence better to bypass frame skip logic
if StopSession is already issued, in order to unblock the pipeline.

Change-Id: I57ffc33816630f7381625296c880917bcf2b42a8