Recorder: Fix stop track issue.
There is a chance, buffer can arrive from camera port
in between EOS is acknowledged by AVCodec and stop is
triggered to Camera context (on kPortIdle CB). If buffer
comes in this scenario then it will be held in queue
without returning back to camera context, as a result
camera context would fail to go in Idle state and stop
track would fail. to fix this problem this commit is
restricting track buffer to go in input list and
returning back immediately if stop is triggred and
EOS is acked by AVCodec.

Change-Id: I5dda4158b65d8e1ebf4b8a8db443a3fa5b55309f