recorder: Fix latency due to race condition in low fps streams
Considering a case in which  current track FPS is 5, and there is a
dynamic FPS update request to 1; supposing next frame was going to be
a valid one, meaning not skipped, the UpdateFPS function would overwrite
remaining_frame_skip_time_, and another 29 frames need to be waited for
next valid frame. This can be avoided, if remaining_frame_skip_time_ is
updated only in FrameSkip logic. Further, this would not affect the case
of high FPS streams, where there is none or rare case of frame skipping.

Change-Id: I4d46ace9af29ba3fb913c06b8e9c2b744a95cb8f