Recorder: MultiCamera: Fix incorrect setting of algo sizes
Private handle have four fields:
width - Aligned width corresponding to stride.
height - Aligned height corresponding to scanline.
unaligned_width - Real width.
unaligned_height - Real height.

Currently to the library was passed width and height
which is stride/scanline and is not correct.

Also lenght was not calculated correctly. Lenght is
size of the plane in bytes which should be: (stride * scnline).

Change-Id: Ib580b91f2784546b991769091001dde6c603f12c