Recorder: MultiCamera: Link related cameras during snapshot
Related cameras are linked during stream creation.
Linking is done by updating the vendor specific metadata
tags. Which are applied on first capture request after
reconfiguration.

We have an issue when first capture image is invoked,
it will cause reconfiguration of the camera
and settings for link related cameras will be lost.

This change reworks the logic and update dual Camera link
metadata tags in capture if reconfigure is done.

Change-Id: I3e64502dfbe4405e412749b3118da852fd3115b4