Recorder: MultiCamera: Stop streams in reverse order
When hw sync is enabled master camera should be stopped last.
The reason for this is that after stopping of master camera
slave camera will also stop sending the frames which causes
instabilities in camera.

Reverse stop sequence is implemented in:
1. Delete streams
2. Stop streams
3. Delete streams

We need reverse order in delete streams as well since
camera context is caching our streams within one session
and they are not stopped/deleted on actual delete call
but only when new stream is created.

Change-Id: I400c9c1e0ef2c24ea5ad447c908203484259372e