Recorder: MultiCamera: Fix frame sync stopping sequence
Request wait and exit is not enough for exiting the
frame Sync thread. It should be send signal as well.
For continues streaming is working since there is
50 ms timeout which is unblocking the condition.

As addition not used function request and exit is removed
since it can not be used for now since have possible race
conditions.

Change-Id: I0f6379fdc71c09dceeb6f053fce6cf899a94dcd1