recorder: MultiCamera: Additional support for multi-sessions
In order to support creating and running a second session
with video track while the first session on the camera is
already running on one instance of MultiCameraManager, all
requests to already created and running streams need to be
temporally halted on CreateStream, so that a new stream can
be added. The same applies for image capture.

Change-Id: Id29939cdd7313fad91228a2a32717b78856d5adf