qmmf-sdk: Change the default Venus Video Encode to Low Power mode
Earlier, the High Quality mode was being set as defualt by the OMX
component for the videos having resolution and fps < (1080p. 60fps)
or Equivalent of this. And Low Power mode was being set as default for
all the videos having reolution and fps higher than the above mentioned.
But this patch sets the Low Power mode for all resolution and fps in
case if the target is Jacala Lite

Change-Id: I30f082df644a7ef194ab8b230b8f83da77aec7a6