recorder: MultiCamera: Fix first dark capture
Camera context is reconfiguring the camera on first
imageCapture. Also there is mechanism for waiting
aec to converge after reconfiguration.
This mechanism is not functional for multicamera usecases,
the reason is that we are linking sessions and
sensors will start after both are in stream on state.

This change:
 - Expose new api for waiting aec to converge
 - Add proper handling for aec converge in multicamera

CRs-Fixed: 2029302
Change-Id: Idc4b1ecfffbd718f169dcca619d3751526eb5b54