recorder: Fix missing Exif data due to code cleanup
The issue was caused by wrong layout of the qmmf_exif_tag_t
structure. That structure must have the same layout as
QEXIF_INFO_DATA.

The filled exif data is passed as pointer to the JPEG
encoder class, which then casts it to a QEXIF_INFO_DATA
and gives it to the encoder library.

Change-Id: Id2729904035062e48c49d9d1fe945a002d9292ff