Recorder: Return track buffers on callback when client died
When client dies track buffers are still sent to the client
via the audio/video callbacks. This causes issues during
cleanup if a session has when more than one video encoded
track.
In order to address this a new flag is defined which will
be raised on DeRegisterClient() when client died. This flag
will cause all buffers received on audio/video callback
to be immediately returned and not sent to the client.

Change-Id: I3959b1c8360ce2b1a03effda2adffb8d0d035d3d