AVCodec: Uneven number of buffers are register and deregister in OpenMax
Issue: For registering the buffer in OMX layer we are using the actual number
of buffers we get from omx client getparameter but while deregistering buffers
we use the number of buffer we allocated not the number omx client returned.

Fix: While deregistering buffer added omx client get parameter to get number of
buffer register and free only those.

Change-Id: I1446b90878f92e1fd29c1170cf655d1e43f28248