Recorder: Optimize snapshot latencies during preview/video.
Dont resume streaming request before submitting the snapshot
request, resume it after submitting snapshot request, this will
prioritize the snapshot request in HAL. this is needed only when
postproc pipe is enabled where streaming request needs to be
paused to avoid multiple stream reconfigurations, this is for
single snapshot case. for continuous snapshot case this change
is avoiding resume before snapshot request and also restoring
the stream ids to streaming request so they can be used in further
update request.

Change-Id: I9e103ad55a1b16386e49854e5618e3f5076fc535