postproc: Fix race condition during delete in HAL reprocess
A race condition between Delete() and AddResult() causes the
HAL reprocess module to get stuck in deadlock during deletion
of the reprocess stream due to both being protected by the same
module_lock_ mutex instead of protecting the deletion of the
stream with reproc_lock_ mutex.
This deadlock caused the flush() of the HAL streams during
Delete() to get stuck because the metadata stream in HAL is not
able to stop as the callback for meta is stuck on AddResult().

Change-Id: I13a70fb063dff4da1fac37ea6f0c5f216ddc6459