qmmf-sdk: Add a new BufferFormat for YUV Encodable buffer
The YUV buffer configured for snapshot is not suitable for
encoding as the memory allocator usage flags set for YUV
buffer is not set to kVideoEncoder.

Add a BufferFormat kNV12Encodable for NV12 YUV buffer which can
be sent to Encoder for encoding. For a buffer of format
kNV12Encodable, right usage flags are set in CreateSnapshotStream.
This will also fix video time lapse ETB failure occured due to
invalid size.

Change-Id: Ic9451795ca6c6722e8f44b72e47d45b4c2c05e12