cameraadaptor : Change GlobalVendorTagDescriptor for Concurrent camera
GlobalVendorTagDescriptor is overwritten for concurrent camera usecase
during StartCamera of 2nd camera. Similarly, CloseCamera of 1st camera
clears the GlobalVendorTagDescriptor, so make it invalid for 2nd camera.

This change solves the above problem by making the vendor_tag_desc_ as
static member and doesn't create the descriptor if it already exists.
Similarly, it doesn't clear the descriptor if other camera is using it.

Change-Id: Ib66a9cb8015b1b46e7801082cb9779155cf2f606