recorder: Add support for multi client support for same camera
Client can open a given camera in slave mode by passing the
kCameraSlaveMode flag in CameraStartParam structure.
The camera being opened as slave needs to be already opened by
another client which uses it as master. If this requirement is
not fulfilled then the slave client needs to wait until signaled
on event from the service.
The slave camera client is required to stop and delete all
sessions and tracks  and call StopCamera API when it receives
EventType::kCameraClosing for the camera which he is using.
Slave camera client has no control over capture image and
get/set camera parameters APIs.

Change-Id: I94698eefe1e4828159f9a76dd482873136c82907