recorder: handle video o/p buffer during flush
When the flush is in progress on the output port
of the encoder then it returns all the queued
buffer immediately without processing it and
all subsequent FTB are also returned immediately.
These buffers are not valid buffers as they don't
have any data in them. Currently, these buffers are
notified to the client. With this change, the
invalid buffers are not notified to the client.
Also, no FTB will be sent to encoder till the
flush is in progress.

Change-Id: I5c465c106a99010c603f2ce8dad3c0961534f516