qmmf-sdk: Add support for exporting ion metadata fd
The ion metadata FD allocated by GBM for each buffer is not
getting propagated to the recorder client process space.

Propagate the metadata FD over to the client via the android
Binder, open an instance of the GBM library and import the
whole buffer so that its info (metadata fd, width, etc.) will
be added to the internal GBM library map. Once added to the
internal map the metadata fd and other info can be extracted
by any other application (in same process space) by importing
only the buffer FD.

Change-Id: Ic86a927df276cf997edb12d19927bd5eb9fc2b9e