qmmf-sdk: Add batch process overlay support
This change adds 2 overlay APIs
- ProcessOverlayItems/Objects
To process overlay objects in a batch. It takes list overlay params
and creates, updates, enables/disables overlay item based on request.

- DeleteOverlayItems/Objects
To delete overlay objects in a batch. It deletes all the overlay items
associated with a track for which this API is called.

Change-Id: I99cc83dc82ba08b62f86c32dd1a16008c3c250d8