recorder: Fix for missing return type in Lambda function
Issue: Sometimes QMMF service is getting died while doing
a callback to client.

Fix: Generally return-type in lambda expression are evaluated by compiler itself
and we don’t need to specify that explicitly and return-type part can be
ignored but in some complex case, compiler can’t make out the return type
and we need to specify that.

Change-Id: I82b0bab281a81393d6a2636395ccf0821ae93690