overlay: Remove downscale logic for bounding box
Currently, if width and height are scaled. For overlay with params
less than the threshold 480x270 a larger buffer is allocated and
for params greater than the threshold smaller buffer is allocated.
This buffer is then scaled to actual param while bliting. Due to
this box with smaller params has shrinked box and box with larger
params has stretched box lines. This change removes the scaling
so that all the boxes appear with approximately uniform thickness.

Change-Id: Ie4a905e0a6c280cf484d08442b58d7af91fff3fd