recorder: track PA client timestamps to data
Tracking the timestamp of the audio buffers
received from PulseAudio to the system clock was
exposing the jitter that is inherent in the
PulseAudio design.  This was causing issues with
downstream components.

Instead, track the audio buffer timestamp to the
data throughput.  The start time is still derived
from the system clock, which helps to keep it
aligned with the camera.

Change-Id: If5b6d74d3105645a7f1f93ff4b4cd29debf983b1