camera-adaptor: Remove the CPU read/write flags in UBWC case
Remove the CPU read/write flags set by HAL since they are
confusing GBM when UBWC flag is set which causes the allocated
buffer to be plain NV12. The GBM library is accepting the UBWC
flag and setting the proper format but then it is overwriten
if CPU write flag is set.

Change-Id: Iad8a7177f26d482626d37f2027ca5873dfbbfd12