recorder: Fix race condition on StopCamera when client is dead
The issue happen when the client calls StopCamera and dies before
StopCamera is completed.

This is happen because when the client dead handler is called,
it calls StopCamera again before the first call is finished.
After that DeInit is called and the camera context is destroyed before
StopCamera is completed. Adding a mutex in StopCamera causes a block in the
second call of StopCamera until the first one finished.

Change-Id: I3e280d9150ff29433fd685f4bf13a165979a09d6