recorder: fix concurrent execution of start/stop session
Issue: StopSession API could be called by client and dead
handler in parallel. This happens if client die while
StopSession is on going. Dead handler also try to call
StopSession on the same time in this scenario.

Fix: Add session mutex to serialize execution of Start/Stop
session. There is already check which protect session to be
started or stopped twice.

Change-Id: Ie432bd5cc6e4a7b37a67cacdf8d8e9d662a4edd5