cameraadaptor: fixes for runtime plug/unplug of cameras
Issue:
- When we disconnect a camera runtime, we set the camera device to
error state which doesn't allow the user to make subsequent calls
in the flow (like stopsession, deletesession, etc) and the server dies.
- We mark the request as pending before submitting, if submit fails we
get request back as ERROR_REQUEST with no input buffer. Due to this
result callback will try to process it resulting crash.
Fix:
- Do not assert in stop sequence, instead do the clean up as usual,
and propagate the error to the client whereever applicable.
Let client handle the error.
- Mark request as pending only when the submit is successful.

Change-Id: I713db96771bd294eccce7fbdcdef9ab1f69e66c1