cameraadaptor: fix marking related issue
Issue: As part of 3870785, sequence of registering and submitting
request was changed. We first submit and then register request if
submit is successful. This is causing issue for scenario where
result is returned before the request is registered.

Fix: We revert back to previous sequence where the request is marked
first and then submitted. The error scenario for submit failure is
handled in error notify where the respective request is removed to
avoid crash in result handling for unplug scenario.

Change-Id: Id925b7e4967d6707dc7bcb4bc0367e35cc61f56c