cameraadaptor: handle buffer coming from camera after unplug
Issue: for the hotplug usecase, after POV unplug, there is
a possibility of successful buffer coming from camera. But
after unplug we set the device state to error state, and
we are not handling this scenario.

Fix: when successful buffer received in ReturnStreamBuffer
after unplug, eventhough in error state added code to handling
this scenario.

Change-Id: Iba480ca89e258e03d630abcff019c1a1e4d2dfaa