qmmf-sdk: fix propagation of GBM buffer FD
According to Mesa/GBM documentation each call to
gbm_bo_get_fd() returns a new file descriptor (FD)
and the caller is responsible for closing the file
descriptor. Therefore we have to close buffer FD
when buffer is released.  But we also want to
retain buffer FD immutable. Because of that we
cache buffer FD in our layer so that GetFD API
return the same FD each time.

Change-Id: Ic82bbc201931ff2e921427ce51c1c5b72b848e0a