qmmf-sdk: add support for sending vendor tag once in the request.
In the video stream, we are doing repeating requests. In which, if
we set some particular vendor tag, then all consecutive requests
will contain the vendor tag as we are reusing the metadata.
But there are some vendor tags, which needs to be sent in the single
request. For the consecutive request we should send the default value of those
vendor tags.

Fix: added support in the qmmf camera context, where we check the
vendor tags, which needs to be sent once, if the metadata has it,
then we send the request once and then we will udpate that vendor
tag with default value and update the streaming_request_.
So that, the consecutive requests will have the default value of
that vendor tag.

Change-Id: Id6ebcf3f6e8567010d4b89d112c7540433ac3514