qmmf-sdk: refactor offlinejpeg and support offlinecamera
Because the offline IPE function and the offline JPEG function
use the same underlying library, but the CamX added new input
parameters to support the offline IPE, it was necessary to refactor
the original offline JPEG framework in qmmf-server. This refactoring
allows the same codebase to support both the offline IPE and offline
JPEG functions. Additionally, for naming consistency, both the offline
IPE and offline JPEG functionalities have been unified under the name
offline camera.

Change-Id: I6cb15119e9752b7f35cfbddf2a12d12b686f1dc8
Signed-off-by: Hui Liu <quic_huliu@quicinc.com>